pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "claim-processing-pipeline"
version = "0.1.0"
description = "FastAPI service for processing PDF insurance claims using a LangGraph multi-agent pipeline"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "langgraph>=0.1.0",
    "langchain>=0.2.0",
    "langchain-openai>=0.1.0",
    "langchain-google-genai>=1.0.0",
    "langchain-core>=0.2.0",
    "pypdf>=4.2.0",
    "pymupdf>=1.24.0",
    "python-multipart>=0.0.9",
    "python-dotenv>=1.0.0",
    "pytesseract>=0.3.10",
    "pdf2image>=1.17.0",
    "Pillow>=10.0.0",
]

[project.optional-dependencies]
dev = [
    "hypothesis>=6.100.0",
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
